vochain/indexer: use sqlite engine in sqlc
It's complete enough that it can parse all of our queries,
so we no longer need to (ab)use the mysql engine and hope for the best.
We update to the latest sqlc master to use a version that works well.

It even solves the issue we were having with LIMIT/OFFSET and args,
since that issue was particular to the mysql parser.

One minor change is that limit/offset parameters are now int64,
which is larger than the old int32. This seems harmless.

Another minor change is that more query parameters are interface{} now.
That seems mildly unfortunate, but it's ultimately also harmless.
If we fed the wrong types into the queries, our tests would blow up.

+++ vochain/indexer/process.go
@@ -123,8 +123,8 @@ func (idx *Indexer) ProcessList(entityID []byte,
 		Status:          int64(statusnum),
 		SourceNetworkID: int64(srcNetworkId),
 		IDSubstr:        searchTerm,
-		Offset:          int32(from),
-		Limit:           int32(max),
+		Offset:          int64(from),
+		Limit:           int64(max),
 		WithResults:     withResults,
 	})
 	if err != nil {
@@ -160,8 +160,8 @@ func (idx *Indexer) EntityList(max, from int, searchTerm string) []types.HexByte
 	// work it out here so that the api package doesn't cause N sql queries.
 	entityIDs, err := idx.readOnlyQuery.SearchEntities(context.TODO(), indexerdb.SearchEntitiesParams{
 		EntityIDSubstr: searchTerm,
-		Offset:         int32(from),
-		Limit:          int32(max),
+		Offset:         int64(from),
+		Limit:          int64(max),
 	})
 	if err != nil {
 		log.Errorf("error listing entities: %v", err)
